% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompBdy.R
\name{CompBdy}
\alias{CompBdy}
\title{Tumor boundary recognition}
\usage{
CompBdy(st_object, NicheLabel = "Mal", minSpot = 10, sliceName = "slices")
}
\arguments{
\item{st_object}{a ST Seurat object}

\item{NicheLabel}{Tumor spot label (default is 'Tumor')}

\item{minSpot}{Remove areas with fewer spots (default is 10)}

\item{sliceName}{Slice name (default is 'slices')}
}
\value{
a ST Seurat object
}
\description{
Tumor boundary recognition
}
