% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompScore.R
\name{CompScore}
\alias{CompScore}
\title{Scoring based on characteristic gene set}
\usage{
CompScore(st_object, assay = "Spatial", slot = "counts", ncores = 1)
}
\arguments{
\item{st_object}{a ST Seurat object}

\item{assay}{default is 'Spatial'}

\item{slot}{Slot to pull expression data from (default is counts)}

\item{ncores}{multi-core}
}
\value{
ST object after scoring
}
\description{
Scoring based on characteristic gene set
}
