% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearSpotA.R
\name{nearSpotA}
\alias{nearSpotA}
\title{Get 6 spots around some spots}
\usage{
nearSpotA(coor_df, spot_all)
}
\arguments{
\item{coor_df}{spatial coordinates data frame}

\item{spot_all}{all spots spatial location}
}
\value{
a data frame
}
\description{
Get 6 spots around some spots
}
