% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearSpotN.R
\name{nearSpotN}
\alias{nearSpotN}
\title{Get 6 spots around a certain spot}
\usage{
nearSpotN(coor_df, spot)
}
\arguments{
\item{coor_df}{spatial coordinates data frame}

\item{spot}{spatial location}
}
\value{
a data frame
}
\description{
Get 6 spots around a certain spot
}
