% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Niche_cat.R
\name{NicheCat}
\alias{NicheCat}
\title{Niche category based on malignancy_score}
\usage{
NicheCat(
  st_object,
  score_column = "malignancy_score",
  niche_column = "Niche",
  formula = "Q1"
)
}
\arguments{
\item{st_object}{a ST Seurat object}

\item{score_column}{malignancy_score column name}

\item{niche_column}{ST Niche column name}

\item{formula}{Choose 'Q1' or 'median'(default is Q1:lower quartile)}
}
\value{
ST object after Niche category
}
\description{
Niche category based on malignancy_score
}
