% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCT.R
\name{SCT}
\alias{SCT}
\title{SCTransform And Clustering}
\usage{
SCT(st_object, assay = "Spatial", npcs = 30, resolution = 0.3)
}
\arguments{
\item{st_object}{a ST Seurat object}

\item{assay}{default is 'Spatial'}

\item{npcs}{Total Number of PCs to compute and store}

\item{resolution}{Value of the resolution parameter}
}
\value{
a normalized ST object
}
\description{
SCTransform And Clustering
}
