% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segSpot.R
\name{segSpot}
\alias{segSpot}
\title{Segment Divide the spot into regions}
\usage{
segSpot(selected_df)
}
\arguments{
\item{selected_df}{Selected spots}
}
\value{
section spot data frame
}
\description{
Segment Divide the spot into regions
}
